feat: add product CRUD with JPA entities, repositories, service and R…

+++ backend/src/main/resources/db/migration/V1__schema.sql
@@ -1,14 +1,14 @@
 CREATE TABLE categories (
-                            id SERIAL PRIMARY KEY,
+                            id BIGSERIAL PRIMARY KEY,
                             name VARCHAR(100) UNIQUE NOT NULL
 );
 
 CREATE TABLE products (
-                          id SERIAL PRIMARY KEY,
+                          id BIGSERIAL PRIMARY KEY,
                           sku VARCHAR(50) UNIQUE NOT NULL,
                           name VARCHAR(200) NOT NULL,
                           description TEXT,
-                          category_id INTEGER REFERENCES categories(id),
+                          category_id BIGINT REFERENCES categories(id),
                           price NUMERIC(10,2) NOT NULL CHECK (price >= 0),
                           stock INTEGER NOT NULL DEFAULT 0 CHECK (stock >= 0),
                           weight_kg NUMERIC(8,3),
@@ -17,16 +17,16 @@ CREATE TABLE products (
 );
 
 CREATE TABLE orders (
-                        id SERIAL PRIMARY KEY,
+                        id BIGSERIAL PRIMARY KEY,
                         status VARCHAR(20) NOT NULL DEFAULT 'completed',
                         total_amount NUMERIC(12,2) NOT NULL,
                         created_at TIMESTAMP DEFAULT NOW()
 );
 
 CREATE TABLE order_items (
-                             id SERIAL PRIMARY KEY,
-                             order_id INTEGER NOT NULL REFERENCES orders(id) ON DELETE CASCADE,
-                             product_id INTEGER NOT NULL REFERENCES products(id),
+                             id BIGSERIAL PRIMARY KEY,
+                             order_id BIGINT NOT NULL REFERENCES orders(id) ON DELETE CASCADE,
+                             product_id BIGINT NOT NULL REFERENCES products(id),
                              quantity INTEGER NOT NULL CHECK (quantity > 0),
                              unit_price NUMERIC(10,2) NOT NULL,
                              subtotal NUMERIC(12,2) NOT NULL
